requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
dj-database-url==1.2.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gspread==5.9.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
kombu==5.2.4
multidict==6.0.4
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.7
outcome==1.2.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.2.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
twilio==8.2.0
uritemplate==4.1.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
wsproto==1.2.0
yarl==1.9.2
